refactor(settings/models): 提取类型定义和工具函数为独立模块
